move logic to prover.rs
